requirements.txt
openai>=1.0.0
langchain-openai>=0.0.5
langchain-community>=0.0.11
streamlit>=1.32.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4 